requirements.txt
alembic==1.15.2
aniso8601==10.0.1
attrs==25.3.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
Deprecated==1.2.18
dotenv==0.9.9
Flask==3.1.1
flask-cors==6.0.0
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.12
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.2
idna==3.10
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
limits==5.2.0
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ordered-set==4.1.0
packaging==25.0
pillow==11.2.1
Pygments==2.19.1
PyJWT==2.10.1
python-dotenv==1.1.0
pytz==2025.2
referencing==0.36.2
requests==2.32.3
rich==13.9.4
rpds-py==0.25.0
SQLAlchemy==2.0.41
stripe==12.1.0
typing_extensions==4.13.2
urllib3==2.4.0
uuid==1.30
Werkzeug==3.1.3
wrapt==1.17.2
